requirements.txt
2captcha-python==1.2.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anticaptchaofficial==1.0.53
anyio==3.7.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
capmonstercloudclient==1.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.6.2
chromedriver-binary==120.0.6061.0.0
chromedriver-binary-auto==0.3.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.1
cssselect==1.2.0
cssutils==2.7.0
cycler==0.11.0
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.1
flake8==6.0.0
fonttools==4.42.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.96.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-vision==3.4.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
gspread==5.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.23.1
ipython==8.14.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.3.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kaitaistruct==0.10
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parsel==1.8.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
playwright==1.34.0
premailer==3.10.0
prompt-toolkit==3.0.38
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
pyairtable==1.5.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
PyDispatcher==2.0.7
pydivert==2.1.0
PyDrive==1.3.1
pyee==9.0.4
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytrends==4.9.2
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
runstats==2.0.0
scraperapi-sdk==0.2.2
Scrapy==2.9.0
scrapy-playwright==0.0.26
scrapy-zyte-api==0.8.4
seaborn==0.12.2
selenium==4.10.0
selenium-wire==5.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.2
tldextract==3.4.4
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
twilio==8.8.0
Twisted==22.10.0
typing_extensions==4.6.3
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.16
w3lib==2.1.1
wcwidth==0.2.6
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
yagmail==0.15.293
yarl==1.9.2
zenrows==1.3.1
zope.interface==6.0
zstandard==0.21.0
zyte-api==0.4.5
